- reverted old cvPoint, silent error is dangerous - fixed bugs in files convexhull.py and contours.py, for python 3 - fixed a bug in file camhisto.py, for python 3 - fixed a python 3 bug in get_pixel_or_slice2d() - fixed a python 3 bug in file motempl.py - fixed a python 3 bug in file squares.py
git-svn-id: http://ctypes-opencv.googlecode.com/svn/trunk@228 7d674749-4f56-0410-b1b8-d5180b90564b

+++ demo/cam-histo.py
@@ -166,7 +166,7 @@ if __name__ == '__main__':
         cvSetZero (histimg)
 
         # compute the width for each bin do display
-        bin_w = histimg.width / hdims
+        bin_w = int(histimg.width / hdims)
         
         for  i in range (hdims):
             # for all the bins
